requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
cryptography
groq
twilio
python-dotenv
pydantic
jinja2
python-multipart
httpx
